fix: RO テスト移植
